1.1.3.8 Beta

+++ scripts/_xkeen/02_install/04_install_geosite.sh
@@ -10,9 +10,11 @@ install_geosite() {
         update_flag=$4
 
         temp_file=$(mktemp)
+        
+        # Первая попытка: прямая загрузка
         if curl -L -o "$temp_file" "$url" > /dev/null 2>&1; then
-            mv "$temp_file" "$geo_dir/$filename"
-            if [ -s "$geo_dir/$filename" ]; then
+            if [ -s "$temp_file" ]; then
+                mv "$temp_file" "$geo_dir/$filename"
                 if [ "$update_flag" = true ]; then
                     echo -e "  $display_name ${green}успешно обновлен${reset}"
                 else
@@ -24,9 +26,25 @@ install_geosite() {
                 return 1
             fi
         else
-            rm -f "$temp_file"
-            echo -e "  ${red}Ошибка${reset} при загрузке $display_name. Проверьте соединение с интернетом или повторите позже"
-            return 1
+            # Вторая попытка: загрузка через прокси
+            if curl -L -o "$temp_file" "$gh_proxy/$url" > /dev/null 2>&1; then
+                if [ -s "$temp_file" ]; then
+                    mv "$temp_file" "$geo_dir/$filename"
+                    if [ "$update_flag" = true ]; then
+                        echo -e "  $display_name ${green}успешно обновлен через прокси${reset}"
+                    else
+                        echo -e "  $display_name ${green}успешно установлен через прокси${reset}"
+                    fi
+                    return 0
+                else
+                    echo -e "  ${red}Неизвестная ошибка${reset} при установке $display_name"
+                    return 1
+                fi
+            else
+                rm -f "$temp_file"
+                echo -e "  ${red}Ошибка${reset} при загрузке $display_name. Проверьте соединение с интернетом или повторите позже"
+                return 1
+            fi
         fi
     }
 
@@ -53,10 +71,10 @@ install_geosite() {
         datfile="geosite_zkeen.dat"
         [ -L "$geo_dir/geosite_zkeen.dat" ] && datfile="zkeen.dat"
         process_geosite_file \
-        "$zkeen_url" \
-        "$datfile" \
-        "GeoSite Zkeen" \
-        "$update_zkeen_geosite"
+            "$zkeen_url" \
+            "$datfile" \
+            "GeoSite Zkeen" \
+            "$update_zkeen_geosite"
         # Создание симлинков для совместимости
         if [ "$datfile" = "geosite_zkeen.dat" ]; then
             rm -f "$geo_dir/zkeen.dat"
